feat(doc-plugin-preview): improve preview
1. support refresh iframe 2. refactor DemoBlock
