+++ .changeset/olive-singers-deny.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/doc-plugin-preview': patch
+---
+
+feat: support refresh iframe and refactor DemoBlock to support dark mode
+feat: 支持刷新 iframe 并且重构 DemoBlock 组件支持暗黑模式
